add detail/parsing/input_adapters.hpp
